requirements.txt
fastapi==0.115.2
uvicorn==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.0